feat(DATAGO-110888): Add dynamic tools into the 'python' tool type (#281)
This feature adds the ability to define 'dynamic' tools as part of the python tool type. For these tools, the code in the tool is able to define its own name, description and parameters programmatically at startup. This allows it to use configuration or even things like a database schema to create one or more tools.